requirements.txt
PyNaCl==1.3.0
dnspython==1.16.0
youtube_dl==2021.3.14
colorlog
pyjokes
howdoi
#guesslang==2.0.1
discord.py==1.6.0
beautifulsoup4==4.9.3
praw
buttons
lyricsgenius
youtube_title_parse
selenium==3.141.0
googletrans==3.1.0a0
urbdict